requirements.txt
python-telegram-bot[job-queue]==21.10
requests>=2.31.0
python-dotenv==1.2.1